linux/requirements.txt
# Core ML
tensorflow==2.15.0
lightgbm==4.1.0
xgboost==2.0.3
scikit-learn==1.3.2
pandas==2.1.0
numpy==1.26.0

# API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Utilities
PyYAML==6.0.1
requests==2.31.0
python-dotenv==1.0.0
joblib==1.3.2

# Monitoring
prometheus-client==0.19.0
